+++ src/components/dropdown-item/dropdown-item.tsx
@@ -1,4 +1,5 @@
 import { Component, ComponentInterface, EventEmitter, h, Event, Listen, Host } from '@stencil/core';
+import { isEnterKey, isSpaceKey } from "../../utils/helpers";
 
 @Component({
     tag: 'wcs-dropdown-item',
@@ -10,15 +11,23 @@ export class DropdownItem implements ComponentInterface {
     })
     wcsDropdownItemClick: EventEmitter<void>;
 
+
     @Listen('mousedown')
     onMouseDown(_: MouseEvent): void {
         this.wcsDropdownItemClick.emit();
     }
 
+    @Listen('keydown')
+    onKeyDown(evt: KeyboardEvent): void {
+        if (isSpaceKey(evt) || isEnterKey(evt)) {
+            this.wcsDropdownItemClick.emit();
+        }
+    }
+
     render() {
         return (
-            <Host slot="item">
-                <slot />
+            <Host slot="item" tabindex="0">
+                <slot/>
             </Host>
         );
     }